+++ sbin/fsck_ffs/pass1b.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: pass1b.c,v 1.17 2009/10/27 23:59:32 deraadt Exp $	*/
+/*	$OpenBSD: pass1b.c,v 1.18 2013/04/24 13:46:29 deraadt Exp $	*/
 /*	$NetBSD: pass1b.c,v 1.10 1996/09/23 16:18:37 christos Exp $	*/
 
 /*
@@ -48,8 +48,9 @@ static ino_t info_inumber;
 static int
 pass1b_info(char *buf, size_t buflen)
 {
-	return (snprintf(buf, buflen, "phase 1b, inode %d/%d",
-	    info_inumber, sblock.fs_ipg * sblock.fs_ncg) > 0);
+	return (snprintf(buf, buflen, "phase 1b, inode %llu/%llu",
+	    (unsigned long long)info_inumber,
+	    (unsigned long long)sblock.fs_ipg * sblock.fs_ncg) > 0);
 }
 
 void
